major cleanup, moved win build ops into existing py build ops files
