+++ requirements.txt
@@ -1,5 +1,12 @@
 git+https://github.com/OpenTrons/autobahn-python.git#egg=autobahn-python
 
+pyinstaller==3.1.1
+pyserial==3.1.1
+PyYAML==3.11
+u-msgpack-python==2.1
+resource
+datetime
+
 #argh==0.26.2
 #attrs==16.0.0
 #cbor==1.0.0
@@ -19,13 +26,9 @@ git+https://github.com/OpenTrons/autobahn-python.git#egg=autobahn-python
 #pyasn1-modules==0.0.8
 #pycparser==2.14
 #Pygments==2.1.3
-pyinstaller==3.1.1
 #PyNaCl==1.0.1
 #pyOpenSSL==16.0.0
-pyserial==3.1.1
 #PyTrie==0.2
-PyYAML
-#==3.11
 #requests==2.10.0
 #sdnotify==0.3.1
 #service-identity==16.0.0
@@ -35,10 +38,8 @@ PyYAML
 #treq==15.1.0
 #Twisted==16.2.0
 #txaio==2.5.1
-u-msgpack-python==2.1
 #watchdog==0.8.3
 #zope.interface==4.2.0
-resource
-datetime
+
 
 # Custom commits/branches
