+++ scripts/zip_ot_app.py
@@ -19,23 +19,32 @@ project_root_dir = \
 electron_app_dir = os.path.join(project_root_dir, "out")
 
 
-def get_build_tag():
+def get_build_tag(os_type):
     arch_time_stamp = "{}{}_{}".format(
         platform.system(),
         struct.calcsize('P') * 8,
         time.strftime("%Y-%m-%d_%H.%M")
     )
 
-    print(script_tag + "Checking Travis-CI environment variables for tag:")
-    travis_tag = tag_from_ci_env_vars(
-        ci_name='Travis-CI',
-        pull_request_var='TRAVIS_PULL_REQUEST',
-        branch_var='TRAVIS_BRANCH',
-        commit_var='TRAVIS_COMMIT'
-    )
+    if os_type == "Linux" or os_type == "Darwin":
+        print(script_tag + "Checking Travis-CI environment variables for tag:")
+        travis_tag = tag_from_ci_env_vars(
+            ci_name='Travis-CI',
+            pull_request_var='TRAVIS_PULL_REQUEST',
+            branch_var='TRAVIS_BRANCH',
+            commit_var='TRAVIS_COMMIT'
+        )
 
-    if travis_tag:
-        return "{}_{}".format(arch_time_stamp, travis_tag)
+        if travis_tag:
+            return "{}_{}".format(arch_time_stamp, travis_tag)
+    if os_type == "Windows":
+        print(script_tag + "Checking Appveyor-CI enironment variables for tag:")
+        appveyor_tag = tag_from_ci_env_vars(
+            ci_name='Appveyor-CI',
+            pull_request_var='APPVEYOR_PULL_REQUEST_NUMBER',
+            branch_var='APPVEYOR_REPO_BRANCH',
+            commit_var='APPVEYOR_REPO_COMMIT'
+        )
 
     return arch_time_stamp
 
@@ -73,8 +82,8 @@ def tag_from_ci_env_vars(ci_name, pull_request_var, branch_var, commit_var):
     return None
 
 
-def zip_ot_app(build_tag):
-    print(script_tab + "Zipping OT App. Using tag: {}".format(build_tag))
+def zip_ot_app(build_tag, os_type):
+    print(script_tab + "Zipping OT App on {}. Using tag: {}".format(os_type,build_tag))
 
     current_app_name = os.listdir(electron_app_dir)[0]
     current_app_path = os.path.join(electron_app_dir, current_app_name)
@@ -85,20 +94,66 @@ def zip_ot_app(build_tag):
     zip_app_dir = os.path.join(project_root_dir, 'releases')
     zip_app_path = os.path.join(zip_app_dir, "opentrons_{}".format(build_tag))
 
-    zip_process = subprocess.Popen(
-        ['zip', '-r', zip_app_path, current_app_path],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE
-    )
+    if os_type == "Linux" or os_type == "Darwin":
+        zip_process = subprocess.Popen(
+            ['zip', '-r', zip_app_path, current_app_path],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE
+        )
+    elif os_type == "Windows":
+        zip_command = 'powershell.exe -nologo -noprofile -command \
+        "& { Add-Type -A \'System.IO.Compression.FileSystem\'; \
+        [IO.Compression.ZipFile]::CreateFromDirectory(\''+current_app_path+\
+        '\',\''+zip_app_path+'\'); }"'
+        zip_process = subprocess.Popen(
+            zip_command,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE
+        )
 
     _, std_err = zip_process.communicate()
     if std_err:
         print(script_tab + "Error using zip command: {}".format(std_err))
 
+def get_os():
+    """
+    Gets the OS to based on the command line argument of the platform info.
+    Only possibilities are: "windows", "mac", "linux"
+    """
+    valid_os = ["windows", "linux", "mac"]
+
+    print(script_tab + "Checking for command line argument indicated OS:")
+    if len(sys.argv) > 1:
+        if sys.argv[1] in valid_os:
+            # Take the first argument and use it as the os
+            print(script_tab + "Valid command line argument found: %s" %
+                  sys.argv[1])
+            return "%s" % sys.argv[1]
+        else:
+            print(script_tab + "Invalid command line argument found: %s\n" %
+                  sys.argv[1] + script_tab + "Options available: %s" % valid_os)
+
+    print(script_tab + "Valid command line arg not found, checking system.")
+
+    os_found = platform.system()
+    if os_found == "Windows":
+        print(script_tab + "OS found is: %s" % valid_os[0])
+        return valid_os[0]
+    elif os_found == "Linux":
+        print(script_tab + "OS found is: %s" % valid_os[1])
+        return valid_os[1]
+    elif os_found == "Darwin":
+        print(script_tab + "OS found is: %s" % valid_os[2])
+        return valid_os[2]
+    else:
+        raise SystemExit("Exit: OS data found is invalid '%s'" % os_found)
 
 def main():
-    print(script_tag + "Zipping OT App")
-    build_tag = get_build_tag()
+    print(script_tag + "Zip procedure started.")
+    print(script_tag + "Checking for OS.")
+    os_type = get_os()
+    print(script_tag + "Zipping OT App for %s." % os_type)
+    build_tag = get_build_tag(os_type)
     zip_ot_app(build_tag)
 
 if __name__ == '__main__':
